+++ src/test/java/com/microsoft/sqlserver/jdbc/connection/XADataSourceTest.java
@@ -5,14 +5,6 @@
 
 package com.microsoft.sqlserver.jdbc.connection;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.security.KeyStore;
-import java.security.cert.CertificateFactory;
-import java.util.ArrayList;
-import java.util.List;
-
 import javax.sql.XAConnection;
 
 import org.junit.jupiter.api.Tag;
@@ -26,13 +18,15 @@ import com.microsoft.sqlserver.testframework.Constants;
 
 
 @RunWith(JUnitPlatform.class)
-@Tag(Constants.XA)
+@Tag(Constants.reqExternalSetup)
 public class XADataSourceTest extends AbstractTest {
     private static String connectionUrlSSL = connectionString + "encrypt=true;trustServerCertificate=false;";
-    private static List<String> certificates = new ArrayList<>();
 
     /**
      * Tests XA connection with PKCS12 truststore that is password protected.
+     * 
+     * Only re-populate the truststore if need arises in the future.
+     * TestUtils.createTrustStore() can be used to create the truststore.
      * 
      * @throws Exception
      */
@@ -40,55 +34,11 @@ public class XADataSourceTest extends AbstractTest {
     public void testPKCS12() throws Exception {
         SQLServerXADataSource ds = new SQLServerXADataSource();
 
-        // populate certificates arraylist with certificates
-        // Only re-populate the truststore if need arises in the future.
-
-        // populateCertificates();
-        // String trustStore = (new TrustStore(certificates)).getFileName();
-
         String trustStore = System.getProperty("pkcs12_truststore");
         String url = connectionUrlSSL + "trustStore=" + trustStore + ";";
         ds.setURL(url);
         ds.setTrustStorePassword(System.getProperty("pkcs12_truststore_password"));
         XAConnection connection = ds.getXAConnection();
         connection.close();
-    }
-
-    private static void populateCertificates() {
-        // populate the arraylist with all the certificates of servers that are used
-        certificates.add("<server name1>.cer");
-        certificates.add("<server name2>.cer");
-    }
-
-    private static class TrustStore {
-        private File trustStoreFile;
-
-        static final String TRUST_STORE_PWD = "<your_password_here>";
-
-        TrustStore(List<String> certificateNames) throws Exception {
-            trustStoreFile = File.createTempFile("<your_truststore_name_here>", null, new File("."));
-            // trustStoreFile.deleteOnExit();
-            KeyStore ks = KeyStore.getInstance("PKCS12");
-            ks.load(null, null);
-
-            for (String certificateName : certificateNames) {
-                ks.setCertificateEntry(certificateName, getCertificate(certificateName));
-            }
-
-            FileOutputStream os = new FileOutputStream(trustStoreFile);
-            ks.store(os, TRUST_STORE_PWD.toCharArray());
-            os.flush();
-            os.close();
-        }
-
-        final String getFileName() throws Exception {
-            return trustStoreFile.getCanonicalPath();
-        }
-
-        private static java.security.cert.Certificate getCertificate(String certname) throws Exception {
-            FileInputStream is = new FileInputStream(certname);
-            CertificateFactory cf = CertificateFactory.getInstance("X.509");
-            return cf.generateCertificate(is);
-        }
     }
 }
